Increase max body size to support larger uploads.
We need a bit more head room to upload our RV/SMR import CSV files which
can be a bit large. Increasing this seems like the simplest solution for
now but if we find we need to support even bigger files in future then
we should probably consider implementing Rails' client-side uploads.